Moved tabs to LayerViewTabs
